use custom rename score during --follow
If you provide a custom rename score on the command line,
like:

  git log -M50 --follow foo.c

it is completely ignored, and there is no way to --follow
with a looser rename score. Instead, let's use the same
rename score that will be used for generating diffs. This is
convenient, and mirrors what we do with the break-score.

You can see an example of it being useful in git.git:

  $ git log --oneline --summary --follow \
	    Documentation/technical/api-string-list.txt
  2dddf1f string-list: Add API to remove an item from an unsorted list
  ecb9149 string_list: Fix argument order for string_list_append
  a82d4d4 string-list: add unsorted_string_list_lookup()
  43d5d6e Fix two leftovers from path_list->string_list
  df6e3f0 Rename path_list to string_list
   create mode 100644 Documentation/technical/api-string-list.txt

  $ git log --oneline --summary -M40 --follow \
	  Documentation/technical/api-string-list.txt
  2dddf1f string-list: Add API to remove an item from an unsorted list
  ecb9149 string_list: Fix argument order for string_list_append
  a82d4d4 string-list: add unsorted_string_list_lookup()
  43d5d6e Fix two leftovers from path_list->string_list
  df6e3f0 Rename path_list to string_list
   rename Documentation/technical/{api-path-list.txt => api-string-list.txt} (47%)
  94fb901 path-list documentation: document all functions and data structures
  e1b3def Start preparing the API documents.
   create mode 100644 Documentation/technical/api-path-list.txt

You could have two separate rename scores, one for following
and one for diff. But almost nobody is going to want that,
and it would just be unnecessarily confusing. Besides which,
we re-use the diff results from try_to_follow_renames for
the actual diff output, which means having them as separate
scores is actively wrong. E.g., with the current code, you
get:

  $ git log --oneline --diff-filter=R --name-status \
            -M90 --follow git.spec.in
  d898118 GIT 0.99.9j aka 1.0rc3
  R084    git-core.spec.in        git.spec.in
  d2add35 Rename the RPM from "git" to "git-core"
  R098    git.spec.in     git-core.spec.in

The first one should not be considered a rename by the -M
score we gave, but we print it anyway, since we blindly
re-use the diff information from the follow (which uses the
default score). So this could also be considered simply a
bug-fix, as with the current code "-M" is completely ignored
when using "--follow".

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>